Mail_Merge/requirements.txt
flask>=3.0.0
flask-wtf>=1.2.1
typing>=3.7.4
python-dotenv>=1.0.0
werkzeug>=3.0.0
email_validator>=2.1.0
dnspython>=2.4.2